move builder to avoid circular imports
